Work around a bug in set_configuration in the Linux xHCI driver.
Linux handles set_configuration requests with an already active
configuration value differently to avoid the costs of a full
reconfiguration. This causes the DATA toggle to desynchronize
(it is reset to DATA0 on the device but not on the host) because of
a bug in the USB3.0 xHCI controller which makes it impossible to
reset an endpoint which was not halted. Adding a set_configuration(0)
forces a full reconfiguration which removes and then readds the
endpoints therby fixing the problem.

This was confirmed and debugged on bus traces of a USB1.1 full-speed
device obtained with the Saleae Logic logic analyzer and is 100%
reproducible.

References:
http://thread.gmane.org/gmane.linux.usb.general/95676
http://thread.gmane.org/gmane.linux.usb.general/122741/focus=125117
http://thread.gmane.org/gmane.linux.kernel.stable/126890