[kvo] Add a method to force a computed update.
The current computed update model is flawed because it is nondeterministic.
Computed updates may be deferred for an arbitrarily long time. This means
that it is very easy to readout stale values.
This is new method a hack to help in situations where we really need
deterministic updates. A proper method to handle this (non-deferred dirty
marking and opportunistic updates) is in the pipeline.