sepack/notify: do not read pin values on startup
The notify channel does not handle 'r' correctly before the first setdebounce
because of a bug in sepack-fw. The commands get stuck in the input buffer and
use up memory. The bug was fixed in 0eefacd76eb5ffb024db145d8b670b59a54527ae
but this is a pretty safe workaround for old devices.