thread: Fix T.queuecall (observables!) on the main thread.
The queue is executed only after a thread yields which is not guaranteed
to happen anytime soon for callbacks added on the main thread. This would
cause a bug in which all observable notification (including those resuming
threads) which were triggered on the main thread or in libev callbacks were
delayed for an indefinite amount of time (basically until an unrelated
thread gets scheduled and then yields).